♻️ :: 코드 리팩토링
한개의 파일에 한개의 클래스만 가지도록 파일 분리, 로직을 build 밖으로 빼서 메모리 낭비를 줄이고, 코드의 가독성을 높임.
